# internal/api/openapi.yaml
openapi: "3.0.3"
info:
  title: CertWatch API
  description: |
    Lightweight, self-hosted SSL/TLS certificate inventory and expiration monitoring.

    All API endpoints except `/health`, `/api/version`, and the auth endpoints require a JWT token obtained via `POST /api/auth/login`.
    Include it as `Authorization: Bearer <token>` in the request header.

    Rate limiting: auth endpoints (`/api/auth/*`) are limited to 10 requests per minute per IP.
    All other authenticated endpoints have no rate limit.

    Request body size is limited to 1 MB.
  version: "1.0.0"
  contact:
    name: CertWatch
    url: https://github.com/araujofrancisco/certwatch
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      description: |
        Creates a new user account.
        Rate-limited to 10 requests per minute per IP.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
        "400":
          description: Validation error (invalid email, weak password, duplicate email)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                weak-password:
                  summary: Weak password
                  value:
                    error: password must be at least 8 characters
                registration-failed:
                  summary: Duplicate email
                  value:
                    error: registration failed
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login and get a JWT token
      description: |
        Authenticates with email and password.
        Rate-limited to 10 requests per minute per IP.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: invalid email or password
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/domains:
    get:
      tags: [Domains]
      summary: List domains
      description: |
        Returns all domains with optional search and enabled-state filtering.
        Tags are included in each domain object.
      operationId: listDomains
      parameters:
        - $ref: "#/components/parameters/q"
        - $ref: "#/components/parameters/enabled"
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Domain list
          content:
            application/json:
              schema:
                type: object
                properties:
                  domains:
                    type: array
                    items:
                      $ref: "#/components/schemas/Domain"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: failed to list domains

    post:
      tags: [Domains]
      summary: Create a domain
      description: |
        Adds a new domain to monitor.
        The domain is auto-scanned in a background goroutine immediately after creation.
      operationId: createDomain
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDomainRequest"
      responses:
        "201":
          description: Domain created
          content:
            application/json:
              schema:
                type: object
                properties:
                  domain:
                    $ref: "#/components/schemas/Domain"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid-domain:
                  summary: Invalid domain name
                  value:
                    error: 'invalid domain name: "bad domain!"'
                duplicate:
                  summary: Duplicate domain
                  value:
                    error: domain already exists
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/domains/import:
    post:
      tags: [Domains]
      summary: Bulk import domains
      description: |
        Import multiple domains at once. Accepts either an array of objects or plain strings.
        Always returns HTTP 200 — individual results have per-item statuses.
      operationId: importDomains
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkImportRequest"
      responses:
        "200":
          description: Import results (per-item statuses)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkAddResponse"
        "400":
          description: Bad request format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                missing:
                  summary: Missing domains field
                  value:
                    error: missing domains
                bad-format:
                  summary: Invalid format
                  value:
                    error: domains must be an array of objects or strings
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/domains/{id}:
    get:
      tags: [Domains]
      summary: Get a domain by ID
      operationId: getDomain
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/domainId"
      responses:
        "200":
          description: Domain details
          content:
            application/json:
              schema:
                type: object
                properties:
                  domain:
                    $ref: "#/components/schemas/Domain"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Domain not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: domain not found

    put:
      tags: [Domains]
      summary: Update a domain
      description: |
        Updates domain name, description, group, enabled status, and tags.
        If the `tags` field is present (even an empty array), existing tags are replaced.
        If the `tags` field is omitted, existing tags are preserved.
      operationId: updateDomain
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/domainId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDomainRequest"
      responses:
        "200":
          description: Domain updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  domain:
                    $ref: "#/components/schemas/Domain"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Domain not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: domain not found

    delete:
      tags: [Domains]
      summary: Delete a domain
      description: |
        Deletes a domain and cascade-deletes associated certificates and tag associations.
      operationId: deleteDomain
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/domainId"
      responses:
        "200":
          description: Domain deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Domain not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: domain not found
  /api/domains/{id}/scan:
    post:
      tags: [Domains]
      summary: Scan a domain
      description: |
        Enqueues an immediate certificate scan across all protocols in priority order:
        HTTPS (5s) -> CT (10s). First success wins. The scan runs asynchronously
        through the background scan queue; this endpoint returns immediately.
        Poll GET /api/domains/{id}/certificates for results.
      operationId: scanDomain
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/domainId"
      responses:
        "202":
          description: Scan queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: queued
        "400":
          description: Invalid domain ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: invalid domain id
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Domain not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: domain not found
  /api/scanqueue:
    get:
      tags: [Domains]
      summary: Scan queue status
      description: |
        Returns the number of scans currently pending in the queue and in flight.
      operationId: scanQueueStatus
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Queue statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  pending:
                    type: integer
                    description: Scans waiting to be dispatched
                    example: 3
                  inflight:
                    type: integer
                    description: Scans currently being executed
                    example: 2
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/certificates:
    get:
      tags: [Certificates]
      summary: List certificates
      description: |
        Returns all certificates with optional filters.
        Supports search, domain, status, protocol, expiry, and expiration filtering.
      operationId: listCertificates
      parameters:
        - $ref: "#/components/parameters/q"
        - $ref: "#/components/parameters/domain_id"
        - $ref: "#/components/parameters/status"
        - $ref: "#/components/parameters/protocol"
        - $ref: "#/components/parameters/expiring"
        - $ref: "#/components/parameters/expired"
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Certificate list
          content:
            application/json:
              schema:
                type: object
                properties:
                  certificates:
                    type: array
                    items:
                      $ref: "#/components/schemas/Certificate"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: failed to list certificates

    delete:
      tags: [Certificates]
      summary: Purge all error certificates
      description: |
        Deletes all certificates with status "error" across all domains.
      operationId: purgeCertificateErrors
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Purge result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteCountResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: failed to purge certificate errors
  /api/certificates/expired:
    delete:
      tags: [Certificates]
      summary: Purge all expired certificates
      description: |
        Deletes all certificates whose `not_after` is in the past across all domains.
        This runs automatically after each background scan, and is triggered per-domain
        when a certificate renewal is detected. Use this endpoint for manual cleanup.
      operationId: purgeExpiredCertificates
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Purge result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteCountResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: failed to purge expired certificates
  /api/domains/{id}/certificates:
    get:
      tags: [Certificates]
      summary: List certificates for a domain
      description: |
        Returns certificates for a specific domain. All certificate list filters apply.
      operationId: listDomainCertificates
      parameters:
        - $ref: "#/components/parameters/domainId"
        - $ref: "#/components/parameters/q"
        - $ref: "#/components/parameters/status"
        - $ref: "#/components/parameters/protocol"
        - $ref: "#/components/parameters/expiring"
        - $ref: "#/components/parameters/expired"
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Certificate list for domain
          content:
            application/json:
              schema:
                type: object
                properties:
                  certificates:
                    type: array
                    items:
                      $ref: "#/components/schemas/Certificate"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Domain not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: domain not found
  /api/domains/{id}/certificates/errors:
    delete:
      tags: [Certificates]
      summary: Purge error certificates for a domain
      description: |
        Deletes all certificates with status "error" for a specific domain.
      operationId: purgeDomainCertificateErrors
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/domainId"
      responses:
        "200":
          description: Purge result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteCountResponse"
        "400":
          description: Invalid domain ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: invalid domain id
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: failed to purge certificate errors
  /api/domains/{id}/certificates/expired:
    delete:
      tags: [Certificates]
      summary: Purge expired certificates for a domain
      description: |
        Deletes all certificates whose `not_after` is in the past for a specific domain.
      operationId: purgeDomainExpiredCertificates
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/domainId"
      responses:
        "200":
          description: Purge result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteCountResponse"
        "400":
          description: Invalid domain ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: invalid domain id
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: failed to purge expired certificates
  /api/reports/inventory:
    get:
      tags: [Reports]
      summary: Inventory report
      description: |
        Returns a combined inventory of all domains with their latest certificate
        and summary statistics. Sorted ascending by days remaining.
      operationId: inventoryReport
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Inventory report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryReport"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from `POST /api/auth/login`.
        Format: `Authorization: Bearer <token>`
  parameters:
    domainId:
      name: id
      in: path
      required: true
      description: Domain ID
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 1
    q:
      name: q
      in: query
      required: false
      description: Search query (LIKE match against domain/description for domains, subject/issuer for certificates)
      schema:
        type: string
      example: example
    enabled:
      name: enabled
      in: query
      required: false
      description: Filter by enabled state (accepts true/false/1/0)
      schema:
        type: boolean
      example: true
    domain_id:
      name: domain_id
      in: query
      required: false
      description: Filter by domain ID
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 1
    status:
      name: status
      in: query
      required: false
      description: Filter by certificate status
      schema:
        type: string
        enum: [valid, expired, error]
      example: valid
    protocol:
      name: protocol
      in: query
      required: false
      description: Filter by discovery protocol
      schema:
        type: string
        enum: [https, ct, smtp, imap, pop3, ldap, ftp, tls]
      example: https
    expiring:
      name: expiring
      in: query
      required: false
      description: Certificates expiring within N days (excludes expired)
      schema:
        type: integer
        minimum: 1
      example: 30
    expired:
      name: expired
      in: query
      required: false
      description: Only show expired certificates
      schema:
        type: boolean
      example: true
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: unauthorized
    RateLimited:
      description: Too many requests (10 req/min per IP on auth endpoints)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: too many requests
  schemas:
    ErrorResponse:
      description: Standard error response
      type: object
      properties:
        error:
          type: string
      required:
        - error
    User:
      description: User account (password is never serialized)
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: User ID
        email:
          type: string
          format: email
          description: Email address
        name:
          type: string
          description: Display name
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
      required:
        - id
        - email
        - name
        - created_at
        - updated_at
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: Email address (must contain @ and .)
          example: admin@example.com
        password:
          type: string
          minLength: 8
          description: Password (minimum 8 characters)
          example: CHANGE_ME
        name:
          type: string
          description: Display name (optional)
          example: Admin
      required:
        - email
        - password
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: admin@example.com
        password:
          type: string
          example: CHANGE_ME
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT token for Bearer authentication
          example: eyJhbGciOiJIUzI1NiIs...
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    Domain:
      description: Monitored domain with optional tags
      type: object
      properties:
        id:
          type: integer
          format: int64
        domain:
          type: string
          description: Domain name (lowercased, trimmed)
          example: example.com
        description:
          type: string
          maxLength: 500
          description: Optional description (max 500 characters)
          example: My site
        enabled:
          type: boolean
          description: Whether monitoring is active
        group:
          type: string
          maxLength: 100
          description: Optional group name (max 100 characters)
          example: production
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - domain
        - enabled
        - created_at
        - updated_at
    Tag:
      description: Tag with random color assignment on first use
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          description: Tag name (lowercased, trimmed)
          example: tls
        color:
          type: string
          description: Hex color code (randomly assigned from 10-color palette)
          example: "#0d6efd"
        created_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - color
        - created_at
    CreateDomainRequest:
      type: object
      properties:
        domain:
          type: string
          description: Domain name to monitor (will be lowercased and trimmed)
          example: example.com
        description:
          type: string
          maxLength: 500
          description: Optional description (max 500 characters)
          example: My site
        group:
          type: string
          maxLength: 100
          description: Optional group name (max 100 characters)
          example: production
        tags:
          type: array
          items:
            type: string
          description: Optional tag names (will be created if they don't exist)
          example: ["tls", "wildcard"]
      required:
        - domain
    UpdateDomainRequest:
      type: object
      properties:
        domain:
          type: string
          description: Updated domain name
          example: example.org
        description:
          type: string
          maxLength: 500
          description: Updated description (max 500 characters)
        group:
          type: string
          maxLength: 100
          description: Updated group name (max 100 characters)
        enabled:
          type: boolean
          description: Updated enabled state
        tags:
          type: array
          items:
            type: string
          description: |
            Replaces all existing tags if present.
            If omitted, existing tags are preserved.
          example: ["updated-tag"]
      required:
        - domain
        - enabled
    BulkImportRequest:
      type: object
      description: Accepts either an array of objects or an array of plain strings
      properties:
        domains:
          oneOf:
            - type: array
              items:
                $ref: "#/components/schemas/BulkDomainObject"
            - type: array
              items:
                type: string
              example: ["example.com", "example.org"]
      required:
        - domains
    BulkDomainObject:
      type: object
      properties:
        domain:
          type: string
          description: Domain name
          example: example.com
        description:
          type: string
          maxLength: 500
          description: Optional description
        group:
          type: string
          maxLength: 100
          description: Optional group name
        tags:
          type: array
          items:
            type: string
      required:
        - domain
    BulkAddResult:
      type: object
      properties:
        domain:
          type: string
          description: Domain name
        status:
          type: string
          enum: [created, skipped, error]
          description: Import status
        error:
          type: string
          description: Error message (only on skipped/error status)
        description:
          type: string
        group:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
        - domain
        - status
...
    BulkAddResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/BulkAddResult"
        summary:
          $ref: "#/components/schemas/BulkAddSummary"
      required:
        - results
        - summary
    Certificate:
      type: object
      description: Discovered SSL/TLS certificate
      properties:
        id:
          type: integer
          format: int64
        domain_id:
          type: integer
          format: int64
        issuer:
          type: string
          description: Certificate issuer distinguished name
          example: "CN=CA, O=Let's Encrypt"
        subject:
          type: string
          description: Certificate subject distinguished name
          example: "CN=example.com"
        serial:
          type: string
          description: Certificate serial number (hex)
          example: "04:AB:CD:EF"
        not_before:
          type: string
          format: date-time
          description: Certificate validity start
        not_after:
          type: string
          format: date-time
          description: Certificate expiry date
        fingerprint:
          type: string
          description: SHA-256 fingerprint
          example: "SHA256:abc123..."
        protocol:
          type: string
          enum: [https, ct, smtp, imap, pop3, ldap, ftp, tls, unknown]
          description: Discovery protocol
        status:
          type: string
          enum: [valid, expired, error]
          description: Certificate status
        sans:
          type: array
          items:
            type: string
          description: DNS names covered by the certificate
          example: ["example.com", "www.example.com"]
        last_checked:
          type: string
          format: date-time
          description: Last scan timestamp
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - domain_id
        - protocol
        - status
        - last_checked
        - created_at
        - updated_at
    DeleteResponse:
      type: object
      properties:
        status:
          type: string
          example: deleted
      required:
        - status
    DeleteCountResponse:
      type: object
      properties:
        deleted:
          type: integer
          description: Number of deleted certificates
          example: 3
      required:
        - deleted
    InventoryReport:
      type: object
      properties:
        inventory:
          type: array
          items:
            $ref: "#/components/schemas/InventoryEntry"
        summary:
          $ref: "#/components/schemas/SummaryStats"
      required:
        - inventory
        - summary
...